Wire OrdersListToolHandler and OrdersGetToolHandler into Hilt multibinding
