+++ libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/di/AiAssistantModule.kt
@@ -7,8 +7,8 @@ import com.woocommerce.android.aiassistant.core.loop.AgenticLoop
 import com.woocommerce.android.aiassistant.core.loop.AgenticLoopImpl
 import com.woocommerce.android.aiassistant.core.loop.ConservativeRetryPolicy
 import com.woocommerce.android.aiassistant.core.loop.HistoryBudgeter
-import com.woocommerce.android.aiassistant.core.loop.RetryPolicy
 import com.woocommerce.android.aiassistant.core.loop.SlidingWindowHistoryBudgeter
+import com.woocommerce.android.aiassistant.core.loop.RetryPolicy
 import com.woocommerce.android.aiassistant.core.loop.ToolCatalogSelector
 import com.woocommerce.android.aiassistant.tools.DefaultToolCatalogSelector
 import com.woocommerce.android.aiassistant.tools.WooCommerceToolRegistry
@@ -16,8 +16,6 @@ import com.woocommerce.android.aiassistant.tools.handlers.AnalyticsOrdersToolHan
 import com.woocommerce.android.aiassistant.tools.handlers.AnalyticsRevenueToolHandler
 import com.woocommerce.android.aiassistant.tools.handlers.CustomersListToolHandler
 import com.woocommerce.android.aiassistant.tools.handlers.OrdersBulkUpdateToolHandler
-import com.woocommerce.android.aiassistant.tools.handlers.OrdersGetToolHandler
-import com.woocommerce.android.aiassistant.tools.handlers.OrdersListToolHandler
 import com.woocommerce.android.aiassistant.tools.handlers.OrdersUpdateToolHandler
 import com.woocommerce.android.aiassistant.tools.handlers.ProductVariationsListToolHandler
 import com.woocommerce.android.aiassistant.tools.handlers.ProductsBulkUpdateToolHandler
@@ -37,63 +35,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 internal abstract class AiAssistantModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindToolRegistry(impl: WooCommerceToolRegistry): ToolRegistry
-
-    @Binds
-    @IntoSet
-    abstract fun bindOrdersListHandler(impl: OrdersListToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindOrdersGetHandler(impl: OrdersGetToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindOrdersUpdateHandler(impl: OrdersUpdateToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindOrdersBulkUpdateHandler(impl: OrdersBulkUpdateToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindProductsListHandler(impl: ProductsListToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindProductsGetHandler(impl: ProductsGetToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindProductsUpdateHandler(impl: ProductsUpdateToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindProductsBulkUpdateHandler(impl: ProductsBulkUpdateToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindProductVariationsListHandler(impl: ProductVariationsListToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindAnalyticsRevenueHandler(impl: AnalyticsRevenueToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindAnalyticsOrdersHandler(impl: AnalyticsOrdersToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindShowCardsHandler(impl: ShowCardsToolHandler): AssistantToolHandler
-
-    @Binds
-    @IntoSet
-    abstract fun bindCustomersListHandler(impl: CustomersListToolHandler): AssistantToolHandler
-
     companion object {
         @Provides
         @Singleton
@@ -101,11 +42,8 @@ internal abstract class AiAssistantModule {
         fun provideAiAssistantJson(): Json = Json {
             ignoreUnknownKeys = true
             encodeDefaults = false
+            explicitNulls = true
         }
-
-        @Provides
-        @Singleton
-        fun provideHistoryBudgeter(): HistoryBudgeter = SlidingWindowHistoryBudgeter()
 
         @Provides
         @Singleton
@@ -124,5 +62,9 @@ internal abstract class AiAssistantModule {
         @Provides
         @Singleton
         fun provideRetryPolicy(): RetryPolicy = ConservativeRetryPolicy
+
+        @Provides
+        @Singleton
+        fun provideHistoryBudgeter(): HistoryBudgeter = SlidingWindowHistoryBudgeter(windowSize = 10)
     }
 }
