+++ libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/di/AiAssistantToolsModule.kt
@@ -1,0 +1,84 @@
+package com.woocommerce.android.aiassistant.di
+
+import com.woocommerce.android.aiassistant.core.chat.AssistantToolHandler
+import com.woocommerce.android.aiassistant.core.chat.ToolRegistry
+import com.woocommerce.android.aiassistant.tools.WooCommerceToolRegistry
+import com.woocommerce.android.aiassistant.tools.handlers.AnalyticsOrdersToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.AnalyticsRevenueToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.CustomersListToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.OrdersBulkUpdateToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.OrdersUpdateToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.ProductVariationsListToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.ProductsBulkUpdateToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.ProductsGetToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.ProductsListToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.ProductsUpdateToolHandler
+import com.woocommerce.android.aiassistant.tools.handlers.ShowCardsToolHandler
+import com.woocommerce.android.aiassistant.tools.orders.OrdersGetToolHandler
+import com.woocommerce.android.aiassistant.tools.orders.OrdersListToolHandler
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import dagger.multibindings.IntoSet
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal abstract class AiAssistantToolsModule {
+    @Binds
+    @IntoSet
+    abstract fun bindOrdersListHandler(handler: OrdersListToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindOrdersGetHandler(handler: OrdersGetToolHandler): AssistantToolHandler
+
+    @Binds
+    @Singleton
+    abstract fun bindToolRegistry(impl: WooCommerceToolRegistry): ToolRegistry
+
+    @Binds
+    @IntoSet
+    abstract fun bindOrdersUpdateHandler(impl: OrdersUpdateToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindOrdersBulkUpdateHandler(impl: OrdersBulkUpdateToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindProductsListHandler(impl: ProductsListToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindProductsGetHandler(impl: ProductsGetToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindProductsUpdateHandler(impl: ProductsUpdateToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindProductsBulkUpdateHandler(impl: ProductsBulkUpdateToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindProductVariationsListHandler(impl: ProductVariationsListToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindAnalyticsRevenueHandler(impl: AnalyticsRevenueToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindAnalyticsOrdersHandler(impl: AnalyticsOrdersToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindShowCardsHandler(impl: ShowCardsToolHandler): AssistantToolHandler
+
+    @Binds
+    @IntoSet
+    abstract fun bindCustomersListHandler(impl: CustomersListToolHandler): AssistantToolHandler
+}
